preliminary support for viewing a player's assignment from coach view
